Clean up the Test Client.
The intent for this should be to do a minimal connectivity test in a way
that is easy for a non-expert user to easily use.  We certainly do not
want it to CREATE new TABLEs and modify an existing Hive instance.

As the first step, we have moved all the connectivity parameters as well
as the query up to the top while cleaning up whitespace and indentation.

We have also removed all the mutating operations so this should only do
a minimal connect and retrieve the list of databases from the Hive
server.